requirements.txt

python-telegram-bot>=20.0
pandas>=1.3.0
xlsxwriter>=3.0.0
